+++ typescript/ui/__tests__/integration/intergation.test.tsx
@@ -18,12 +18,11 @@ import {
   fireEvent,
 } from "@testing-library/react";
 import {
-  ChatResponseType,
-  MessageRole,
   RagbitsClient,
   RagbitsContextProvider,
   StreamCallbacks,
   ChatResponse,
+  ChatResponseType,
 } from "@ragbits/api-client-react";
 import { useConfigContext } from "../../src/core/contexts/ConfigContext/useConfigContext";
 import { ConfigContextProvider } from "../../src/core/contexts/ConfigContext/ConfigContextProvider";
@@ -55,6 +54,16 @@ const historyStore = createStore(createHistoryStore);
 (useHistoryStore as Mock).mockImplementation(
   (selector: (s: HistoryStore) => unknown) => selector(historyStore.getState()),
 );
+
+const chatResponseTypes: ChatResponseType[] = [
+  "text",
+  "reference",
+  "state_update",
+  "message_id",
+  "conversation_id",
+  "live_update",
+  "followup_messages",
+] as const;
 
 describe("Integration tests", () => {
   const BASE_URL = "http://127.0.0.1:8000";
@@ -94,18 +103,16 @@ describe("Integration tests", () => {
       expect(config).toHaveProperty("feedback");
 
       expect(config.feedback).toHaveProperty("like");
-      expect(config.feedback.like).toHaveProperty("enabled");
       expect(typeof config.feedback.like.enabled === "boolean").toBe(true);
-      expect(config.feedback.like).toHaveProperty("form");
+      expect(config.feedback).toHaveProperty("like");
       expect(
         config.feedback.like.form === null ||
           config.feedback.like.form instanceof Object,
       ).toBe(true);
 
       expect(config.feedback).toHaveProperty("dislike");
-      expect(config.feedback.dislike).toHaveProperty("enabled");
       expect(typeof config.feedback.dislike.enabled === "boolean").toBe(true);
-      expect(config.feedback.dislike).toHaveProperty("form");
+      expect(config.feedback).toHaveProperty("dislike");
       expect(
         config.feedback.dislike.form === null ||
           config.feedback.dislike.form instanceof Object,
@@ -169,9 +176,9 @@ describe("Integration tests", () => {
             history: [
               {
                 content: "Test message",
-                role: MessageRole.USER,
+                role: "user",
               },
-              { content: expect.any(String), role: MessageRole.ASSISTANT },
+              { content: expect.any(String), role: "assistant" },
             ],
             message: "Test message 2",
           },
@@ -244,11 +251,11 @@ describe("Integration tests", () => {
             history: [
               {
                 content: "Test message",
-                role: MessageRole.USER,
+                role: "user",
               },
-              { content: expect.any(String), role: MessageRole.ASSISTANT },
-              { content: "Test message 2", role: MessageRole.USER },
-              { content: expect.any(String), role: MessageRole.ASSISTANT },
+              { content: expect.any(String), role: "assistant" },
+              { content: "Test message 2", role: "user" },
+              { content: expect.any(String), role: "assistant" },
             ],
             message: "Test message 3",
           },
@@ -273,7 +280,7 @@ describe("Integration tests", () => {
           const modifiedCallbacks = {
             ...(callbacks as StreamCallbacks<unknown>),
             onMessage: (event: ChatResponse) => {
-              expect(event.type).toBeOneOf(Object.values(ChatResponseType));
+              expect(event.type).toBeOneOf(chatResponseTypes);
             },
           };
 
@@ -308,7 +315,7 @@ describe("Integration tests", () => {
       beforeEach(() => {
         messageId = historyStore.getState().primitives.addMessage({
           content: "Mock content",
-          role: MessageRole.ASSISTANT,
+          role: "assistant",
           serverId: "msg-123",
         });
       });